fix: поправил удоление карточки

+++ controllers/cards.js
@@ -29,18 +29,22 @@ const createCard = (req, res) => {
 const deleteCard = (req, res) => {
   const { cardId } = req.params;
   Card.findById(cardId).then((card) => {
-    if (String(card.owner) === String(req.user._id)) {
-      Card.findByIdAndRemove(cardId)
-        .then(() => {
-          res.send({ message: 'Карточка удалена' });
-        })
-        .catch((e) => {
-          if (e.status === 404) {
-            res.status(404).send({ message: 'Карточка не найдена' });
-          } else {
-            res.status(500).send({ message: 'Smth went wrong' });
-          }
-        });
+    if (card) {
+      if (String(card.owner) === String(req.user._id)) {
+        Card.findByIdAndRemove(cardId)
+          .then(() => {
+            res.send({ message: 'Карточка удалена' });
+          })
+          .catch((e) => {
+            if (e.name === 'CastError') {
+              res.status(400).send({ message: 'Карточка не найдена' });
+            } else {
+              res.status(500).send({ message: 'Smth went wrong' });
+            }
+          });
+      }
+    } else {
+      res.status(404).send({ message: 'Карточка не найдена' });
     }
   });
 };
